feat: issue report form becomes a proper dialog (JUN-213)
Feedback pass: the report form moves from a composer-anchored popover to
the shared Dialog primitive (portal, scrim, focus trap, Esc), the
description field uses the house dialog-textarea (neutral focus ring, no
brand tint), and the footer reads ghost "Add files" left, solid "Send
report" right. Drag-drop targets the whole form body.
